Implement annotating storage and retrieval

+++ src/main/java/ru/turbogoose/cca/backend/model/Dataset.java
@@ -24,7 +24,7 @@ public class Dataset {
     @Column(nullable = false)
     private Long size;
     @Column(nullable = false)
-    private Long totalRows;
+    private Integer totalRows; // TODO: Long?
     @Column(nullable = false)
     private LocalDateTime created;
     private LocalDateTime lastUpdated;
